refactor: split loop helpers
