+++ src/utils/mod.rs
@@ -7,4 +7,7 @@ pub use decryptor_helper::decrypt_full_stream;
 mod md5;
 pub use self::md5::md5;
 
-pub(crate) mod loop_iterator;
+mod loop_counter;
+mod loop_iterator;
+pub use loop_counter::LoopCounter;
+pub use loop_iterator::LoopIter;
